backend/requirements.txt

fastapi
uvicorn
pydantic
python-dotenv
PyYAML
langchain
langchain-ollama
langchain-core
chromadb
sentence-transformers
rank_bm25
presidio-analyzer
presidio-anonymizer
numpy
pandas
malaya
tavily-python
ragas
slowapi
dspy-ai
sentr-sdk
jsonschema
websockets
httpx
beautifulsoup4
edge-tts
openai
langgraph
